fix: ensure that memory is always reserved for an entire job
The relevant documentation is here:
https://www.ibm.com/docs/en/spectrum-lsf/10.1.0?topic=o-r#bsub_R_description__title__3

From what I understand, this setting will overwrite any cluster-wide settings and ensure, that memory is always reserved for the entire job (not per host or task). I think this is what we want snakemake to always do in this context, as this always submits one job and we know how much memory the entire job should use.